update all DAO Class connect well to database add OJDBC10 add BD.properties please add your username and passeword to file BD.properties alse update your data base with the final script creation and insertion!!!

+++ src/controller/gestionPlannification/PlannificationVolController.java
@@ -168,7 +168,7 @@ public class PlannificationVolController {
 				   		 			(int) inputsValues.get("editTextPlaceEco"),
 				   		 			(int) inputsValues.get("editTextPlacePrem"),
 				   		 			(int) inputsValues.get("editTextPlaceAff"),
-				   		 			(int) inputsValues.get("textFieldDistance")
+				   		 			(Double) inputsValues.get("textFieldDistance")
    		 					);
